removed license
The import to Zenodo failed, and the error message indicated that this was due to a metadata import error. So it was likely due to the changes made to the .zenodo.json file. Some Googling indicates that the issue may be with the license entity, so I am removing it to see if it makes a difference.